캐릭터 이동, 델타 타임
1. 키 입력받고 델타 타임으로 캐릭터 이동
2. 델타 타임 계산

+++ zelda-2d/AnimationObject.cpp
@@ -28,7 +28,7 @@ const int AnimationObject::GetAnimationBitmapSize()
 
 const int AnimationObject::GetBitmapCount(const int index)
 {
-	if (0 >= index && index < animationBitmapData.size())
+	if (0 >= index && index < static_cast<int>(animationBitmapData.size()))
 		return bitmapCount[index];
 	else
 		return NULL;
